marwinh-palinda-1/src/twopartsum/twopartsum.go
package main

import (
	"fmt"
)

// sum the numbers in a and send the result on res.
func sum(a []int, res chan<- int) {
	total := 0
	for _, value := range a {
		total += value
	}
	res <- total
}

// concurrently sum the array a.
func ConcurrentSum(a []int) int {
	n := len(a)
	ch := make(chan int)
	go sum(a[:n/2], ch)
	go sum(a[n/2:], ch)
	sum1 := <-ch
	sum2 := <-ch
	totalSum := sum1 + sum2
	return totalSum
}

func main() {
	// example call
	a := []int{1, 2, 3, 4, 5, 6, 7}
	fmt.Println(ConcurrentSum(a))
}
